+++ src/googol/PairClockFlag.java
@@ -1,14 +1,19 @@
 package googol;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
+/**
+ * Classe que guarda um Stamp de comunicação e uma flag para Barrels
+ * Flag: 0 -> AM, unsynchronized; 1 -> AM, synchronized; 2 -> NZ, unsynchronized; 3 -> NZ, synchronized
+ * @author Joao Nuno Coelho
+ * @version 1.0
+ */
 class PairClockFlag{
-		AtomicInteger clock;
-		//0 -> AM, unsynchronized; 1 -> AM, synchronized; 2 -> NZ, unsynchronized; 3 -> NZ, synchronized
+		AtomicLong clock;
 		int flag;
 
 		PairClockFlag(int clock, int flag){
-				this.clock = new AtomicInteger(clock);
+				this.clock = new AtomicLong(clock);
 				this.flag = flag;
 		}
 }
